Rename ConfigActivity.kt
