+++ demo/src/main/res/layout/activity_settings.xml
